requirements.txt
docx-mailmerge == 0.5.0